Refactor(community) : 커뮤니티 관련 엔티티 패키지 변경
