--- src/main/java/kr/co/goalkeeper/api/model/entity/ChattingType.java
@@ -1,5 +1,0 @@
-package kr.co.goalkeeper.api.model.entity;
-
-public enum ChattingType {
-    MESSAGE,URL,IMAGE
-}
