+++ src/main/java/kr/co/goalkeeper/api/model/entity/community/Chatting.java
@@ -1,4 +1,4 @@
-package kr.co.goalkeeper.api.model.entity;
+package kr.co.goalkeeper.api.model.entity.community;
 
 import lombok.Getter;
