+++ src/main/java/kr/co/goalkeeper/api/model/entity/community/ChattingRoomLevel.java
@@ -1,4 +1,4 @@
-package kr.co.goalkeeper.api.model.entity;
+package kr.co.goalkeeper.api.model.entity.community;
 
 public enum ChattingRoomLevel {
     LV1,LV2,LV3,LV4,LV5,LV6,LV7
