+++ src/main/java/kr/co/goalkeeper/api/model/entity/community/ChattingType.java
@@ -1,0 +1,5 @@
+package kr.co.goalkeeper.api.model.entity.community;
+
+public enum ChattingType {
+    MESSAGE,URL,IMAGE
+}
